+++ packages/integrations/cloudflare/test/fixtures/incremental-concurrency/astro.config.mjs
@@ -1,0 +1,22 @@
+import cloudflare from '@astrojs/cloudflare';
+import mdx from '@astrojs/mdx';
+import { defineConfig } from 'astro/config';
+
+export default defineConfig({
+	integrations: [mdx()],
+	adapter: cloudflare({
+		imageService: 'compile',
+	}),
+	output: 'static',
+	image: {
+		service: {
+			entrypoint: './src/image-service.ts',
+		},
+	},
+	build: {
+		concurrency: 2,
+	},
+	experimental: {
+		incrementalBuild: true,
+	},
+});